services/brain/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
torch==2.1.2
prometheus-client==0.19.0
pydantic==2.5.3
python-dotenv==1.0.0
supabase==2.3.7
postgrest==0.16.2
